requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
chromadb>=0.4.22
beautifulsoup4>=4.12.2

# API clients
newsapi-python>=0.2.7

# Data processing
numpy>=1.24.3
pandas>=2.0.3

# Utilities
pytz>=2023.3
python-dateutil>=2.8.2
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.1
flake8>=6.1.0
